+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod dcpu;
 pub mod instructions;
 pub mod assembler;
 pub mod disassembler;
+pub mod devices;